+++ pkg/wellknown/wellknown.go
@@ -1,0 +1,8 @@
+package wellknown
+
+const (
+	ANNOTATION_CONFIGURE_SIDECAR = "greymatter.io/configure-sidecar"
+	ANNOTATION_LAST_APPLIED      = "greymatter.io/last-applied"
+	LABEL_CLUSTER                = "greymatter.io/cluster"
+	LABEL_WORKLOAD               = "greymatter.io/workload"
+)